Send batches of series that match the querier's desired batch size and remove unsafe pooling of chunk slices.
See https://github.com/grafana/mimir/pull/4886#discussion_r1185679925
for discussion of pooling of chunk slices.

+++ pkg/querier/distributor_queryable.go
@@ -54,7 +54,7 @@ type StreamingSeries struct {
 
 type StreamingSeriesSource struct {
 	StreamReader *SeriesChunksStreamReader
-	SeriesIndex  int
+	SeriesIndex  uint64
 }
 
 func newDistributorQueryable(distributor Distributor, iteratorFn chunkIteratorFunc, cfgProvider distributorQueryableConfigProvider, logger log.Logger) QueryableWithFilter {
